python/examples/by-use-case/evm-purchase-on-amazon/requirements.txt
python-dotenv>=1.0.0
openai>=1.7.0
eth_typing>=5.2.1
aiohttp>=3.9.5
base58>=2.1.0
pydantic>=2.0.0
web3>=6.0.0
goat-sdk-wallet-web3>=0.1.0

# Note: readline and asyncio are part of the Python standard library so they don't need installation

# For local development, install goat packages from local directory:
# pip install -e ../../..
# pip install -e ../../../src/wallets/evm
# pip install -e ../../../src/wallets/web3
# pip install -e ../../../src/plugins/crossmint-headless-checkout 